+++ lib/sirius_backend.dart
@@ -11,9 +11,11 @@ export 'src/helpers/file_watcher.dart';
 export 'src/validator.dart';
 export 'src/validation_rules.dart';
 export 'src/wrapper.dart';
-export 'src/web_socket_request.dart';
 
 // TODO: Export any libraries intended for clients of this package.
 
 // apply middleware in only selected methods
+
 // need to add middleware in web sokets also
+
+// Exception shown in production and error logs creation